Move IO tests to IO dir
